refactor: product-post 도메인 모듈 레이어 별로 디렉토리 분리
