+++ apps/api-server/src/product-post/api/dto/find-paged-product-posts-response.dto.ts
