+++ apps/api-server/src/product-post/api/dto/generate-presigned-url-list-response.dto.ts
